commit to save progress, contains lots of testing code added ability to add triggers to mission templates and to add quick strings small refactor of ExtractVector changed fields in operation.h and rgl_string.h to reflect their true (according to me) type

+++ WarbandLib/rgl_string.h
@@ -11,8 +11,8 @@ namespace rgl
 	public:
 		int cursor;
 		char *buffer;
-		size_t buffer_length;
-		size_t str_length;
+		int buffer_length;
+		int str_length;
 		char static_buffer[128];
 
 	public:
